Aura/requirements.txt
colorama
tqdm
alive-progress
requests
aiohttp
beautifulsoup4
fake-useragent
pyjwt
shodan
mmh3
waybackpy
python-whois
dnspython
googlesearch-python
boto3
python3-nmap
ipinfo
ratelimit
PyExecJS
fpdf
pyyaml
